requirements.txt
pandas
numpy
tensorflow-gpu==1.15
opencv-python
awscli
urllib3
tqdm
lxml
tqdm
flask
seaborn
pillow
setuptools
cython
matplotlib
pycocotools
matplotlib
scikit-image
keras
h5py
imgaug
IPython[all]

